test: Reorganize e2e server test suites
Change-type: patch
